Update csscode2.css
Fixed issue which caused scrolling to see full page to be impossible.

Details: 
Having both "box-sizing: border-box" and a fixed height in the body definition caused the browser to always show only a window's worth of text regardless of the size of the body text.
Unfortunately fixing this issue caused another to arise: in pages where body text does not fill a window's worth of space, the background colour now repeats unnecessarily. I will work on this issue next.